requirements.txt
pytest==8.4.2
Appium-Python-Client==5.2.4
selenium==4.35.0